moved functions from exports to main file got server to client recording and transcript working

+++ pi/server.js
@@ -2,18 +2,20 @@ const express = require("express");
 const app = express();
 const SerialPort = require("serialport");
 const port = 3000;
-const opn = require('opn');
+const fs = require('fs');
+const SpeechToTextV1 = require('watson-developer-cloud/speech-to-text/v1');
+//const opn = require('opn');
 const server = require('http').Server(app);
 const io = require('socket.io')(server);
 //const Sys = require('util');
 //const exec = require('child_process').exec;
 const { spawn } = require('child_process');
-var child;//child process variable
+//var child;//child process variable
 //const microphone = require('./mic-rec.js');
 var micState; //Variable to set state of microphone for recording funcitons
 const serial = new SerialPort('/dev/ttyUSB0', {	baudRate: 9600});// Monitor pi usb serial port
 var state;// Variable to hold diary state true = open | false = closed
-
+//const speechText = require('./speech-text.js');
 
 //function requirements
 const mongoDB = require('./push-mongo-DB.js'); //enable access to pushtomongo function
@@ -36,10 +38,18 @@ serial.on('data', (data) => {
 server.listen(5000); // Port for socket to listen on
 
 io.on("connection", (socket) => {
-	setInterval(() => {
-		socket.emit("bookOpen", state);
-	}, 1000);
+    // EMIT TO CLIENT
+    setInterval(() => {
+	socket.emit("bookOpen", state);	
+    }, 1000);
     //Send book status to client every second - where it will be checked and page changed accordingly.
+    
+    onResolveSpeech(socket);
+    
+    //emitOnVarChange(socket, transcript, "transcript", transcript);
+    
+    
+    // BACK FORM CLIENT: 
     socket.on('startRecord', (data) => {
         micState = data;
 	handleRecording();
@@ -53,11 +63,41 @@ io.on("connection", (socket) => {
     });
 
     socket.on('saveEntry', (data) => {
-        console.log("this is when it would save the file and do speech to text");
+	console.log("saving text: ", data);
+	beginAnalysis(data);
     });
+
+
 });
 
+function emitOnVarChange(socket, varName, eventName, data ) {
+    //Funciton that checks a variable on interval and emits variable value when set (only once)
+	var sent = 0;
+	setInterval(() => {
+		if (varName && cnt === 0) {
+		    socket.emit(eventName, data);
+		    sent = 1;
+		}
+	},1000);
+}
 
+function onResolveSpeech(socket) {
+    //This function checks every second until 
+    var cnt = 0;
+	setInterval(() => {
+	    if (transcript && cnt === 0) {
+		socket.emit("transcript", transcript);
+		cnt = 1;
+	    }
+	}, 1000);
+}
+
+
+//END SOCKETS
+
+
+// MICROPHONE RECORDING FUNCTIONS
+var child;
 function startRecording(dir,file) {
   console.log("recording");
 
@@ -77,14 +117,58 @@ function startRecording(dir,file) {
 }
 
 function stopRecording() {
-	console.log('killing');
-	child.kill();
-  var tempMessage = "To start off, I think I completely failed my geometry quiz, which I know I should’ve studied more for...my dad’s not gonna be happy about that. :( Then, we had a pop quiz in history on the reading homework from last night, and I completely forgot most of what I read, which made me even more upset because I actually did the reading! But what really made me mad was the note that Sarah slipped into my locker during passing period. She said she was sad that I’ve been hanging out with Jane more lately and thinks that I don’t want to be her friend anymore. I can’t believe she thinks that, especially after talking with her on the phone for hours and hours last month while she was going through her breakup with Nick! Just because I’ve been hanging out with Jane a little more than usual doesn’t mean I’m not her friend anymore. She completely blew me off at lunch, and when I told Jane, she thought that Sarah was being a";
-  console.log(tempMessage);
-  beginAnalysis(tempMessage);
+    console.log('killing');
+    if (child) child.kill();
+    var text = toText('./', 'test1.wav');
+  //var tempMessage = "To start off, I think I completely failed my geometry quiz, which I know I should’ve studied more for...my dad’s not gonna be happy about that. :( Then, we had a pop quiz in history on the reading homework from last night, and I completely forgot most of what I read, which made me even more upset because I actually did the reading! But what really made me mad was the note that Sarah slipped into my locker during passing period. She said she was sad that I’ve been hanging out with Jane more lately and thinks that I don’t want to be her friend anymore. I can’t believe she thinks that, especially after talking with her on the phone for hours and hours last month while she was going through her breakup with Nick! Just because I’ve been hanging out with Jane a little more than usual doesn’t mean I’m not her friend anymore. She completely blew me off at lunch, and when I told Jane, she thought that Sarah was being a";
+
+  //beginAnalysis(tempMessage);
+}
+
+// END MICROPHONE RECORDING FUNCTIONS
+
+
+
+// SPEECH TO TEXT
+var speechToText = new SpeechToTextV1({
+    username: '4bab5a29-863d-4d9d-8838-3e2dc3ccdcf1',
+    password: 'vnMnoG300QVL',
+    url: 'https://stream.watsonplatform.net/speech-to-text/api'
+  });
+
+
+var params = {
+    objectMode: false,
+    content_type: 'audio/wav',
+    model: 'en-GB_BroadbandModel'
+};
+
+var transcript;
+function toText(dir,file) {
+	// Create the stream.
+	var recognizeStream = speechToText.recognizeUsingWebSocket(params);
+
+	// Pipe in the audio.
+	fs.createReadStream(dir+file).pipe(recognizeStream);
+
+	//print to console
+	recognizeStream.setEncoding('utf8');
+
+
+	// Listen for events.
+	recognizeStream.on('data', function(event) { onEvent('Data:', event); });
+	recognizeStream.on('error', function(event) { onEvent('Error:', event); });
+	//recognizeStream.on('close', function(event) { onEvent('Close:', event); });
+
+	// Display events on the console	
+	function onEvent(name, event) {
+		//console.log("EVENT: ", event);
+		transcript = event;
+	};
 }
 
 
+// END SPEECH TO TEXT
 
 function beginAnalysis(message){
 
@@ -113,13 +197,13 @@ function handleRecording() {
 	//Switch statement to handle recording based on state of book open/closed/stopped etc.
 	switch(micState) {
 	    case "record":
-	      //microphone.startRecord("/home/pi/", "test1.wav");
-		startRecording("/home/pi/", "test1.wav");
+		startRecording("./", "test1.wav");
 	      break;
 	    case "stopped":
 		console.log('stopping the recording?');
-	      //microphone.stopRecord();
 		stopRecording();
+		//console.log(text);
+  //beginAnalysis(tempMessage); don't we want to do this on hit save
 	      break;
 	    default:
 	      console.log('idle');
